few changes
The list page cant refresh and after you have deleted an entry you cant
press the home button to refresh the list, for whatever reason. It has
to do with the way the data is loaded from the .pagebeforeshow, That
might have to be changed or I need to add a pageinit function to idsplay
all hte information.